cli: extract human render helpers ( W4 )
